semana05/dia2/requirements.txt
django==3.2
djangorestframework
django-cors-headers

##para conectar un app de react con django usamos

